feat(runtime): restructure module organization and add query parameter mappers
