+++ src/runtime/server/api/raw-content.get.ts
@@ -1,5 +1,5 @@
 import { defineEventHandler, getQuery, type H3Event } from 'h3';
-import { mapStringToQueryParams } from '../../../mappers';
+import { mapStringToQueryParams } from '../../mappers';
 import { getContentIslandClient } from '../plugins/client';
 
 export default defineEventHandler(async (event: H3Event) => {
